backend/app/requirements.txt
Flask==2.2.0
Flask-RESTX==1.0.3
psycopg2-binary==2.9.6
requests==2.28.1
pandas==1.5.3
werkzeug==2.3.0
